requirements.txt
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
cycler==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
nltk==3.5
numpy==1.20.1
pdfminer.six==20201018
Pillow==8.1.2
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
regex==2020.11.13
six==1.15.0
sortedcontainers==2.3.0
tqdm==4.59.0
wordcloud==1.8.1
